# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger TerraMafia",
        "description": "Api documentation TerraMafia",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "OpenApi host"
        }
    ],
    "paths": {
...
        "/api/v1/me": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Персональные данные текущего пользователя.",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\User\\IUserAccountController::getMe",
                "responses": {
                    "200": {
                        "description": "Персональные данные.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Требуется авторизация."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Account"
                ],
                "summary": "Запрос обновит профиль пользователя",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\User\\IUserAccountController::updatePlayer",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePlayerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Возвращает обновленного пользователя",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Запрос добавления  персональных данных пользователя",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\User\\IUserAccountController::addPersonalInfo",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddPersonalRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Персональные данные успешно добавлены."
                    },
                    "401": {
                        "description": "Требуется авторизация."
                    },
                    "422": {
                        "description": "Ошибка валидации."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/me/email/confirmation": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Подтверждение email пользователя по токену",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\User\\IUserAccountController::confirmEmail",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangeEmailConfirmationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Email успешно подтвержден",
                        "headers": {
                            "Set-Cookie": {
                                "description": "Установка ACCESS_TOKEN cookie",
                                "schema": {
                                    "type": "string",
                                    "example": "token=***; Path=/; HttpOnly"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespondWithToken"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Требуется авторизация"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/me/email": {
            "put": {
                "tags": [
                    "Account"
                ],
                "summary": "Отправка заявки на потверждение email пользователя",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\User\\IUserAccountController::sendEmailVerification",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SendEmailRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Email успешно изменен"
                    },
                    "422": {
                        "description": "Ошибка валидации или пароль введен неверно"
                    }
                }
            }
        },
        "/api/v1/me/password": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Изменение пароля авторизованного пользователя",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\User\\IUserAccountController::changePassword",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangePasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Пароль успешно изменен"
                    },
                    "422": {
                        "description": "Ошибка валидации или пароль введен не верно"
                    },
                    "401": {
                        "description": "Требуется авторизация"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RespondWithToken": {
                "title": "Ответ с токеном аутентификации.",
                "properties": {
                    "message": {
                        "description": "Сообщение.",
                        "type": "string",
                        "example": "Аутентификация прошла успешно."
                    },
                    "ttl": {
                        "description": "Время жизни токена (в минутах).",
                        "type": "integer",
                        "example": "60"
                    },
                    "uuid": {
                        "description": "Глобальный идентификатор пользователя.",
                        "type": "string",
                        "example": "1c6383b4-90b7-431e-a9c0-fb398ea3225f"
                    }
                },
                "type": "object"
            },
            "AddPersonalRequest": {
                "title": "Запрос на добавление персональных данных игрока",
                "required": [
                    "fullName",
                    "gender",
                    "specialisationId"
                ],
                "properties": {
                    "firstName": {
                        "title": "Имя пользователя(игрока)",
                        "description": "Класс валидации запроса на добавление персональных данных игрока.",
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 1,
                        "example": "Федор"
                    },
                    "lastName": {
                        "title": "Фамилия пользователя(игрока)",
                        "description": "Класс валидации запроса на добавление персональных данных игрока.",
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 1,
                        "example": "Панасенко"
                    },
                    "patronymic": {
                        "title": "Отчество пользователя(игрока)",
                        "description": "Класс валидации запроса на добавление персональных данных игрока.",
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 1,
                        "example": "Остапович"
                    },
                    "birthdate": {
                        "title": "Дата рождения",
                        "description": "Класс валидации запроса на добавление персональных данных игрока.",
                        "type": "string",
                        "format": "date",
                        "example": "1981-10-10",
                        "nullable": true
                    },
                    "gender": {
                        "title": "Пол",
                        "description": "Класс валидации запроса на добавление персональных данных игрока.",
                        "type": "string",
                        "enum": [
                            "male",
                            "female",
                            "other"
                        ],
                        "example": "male"
                    }
                },
                "type": "object"
            },
...
            "ChangeEmailConfirmationRequest": {
                "title": "Запрос на подтверждение нового email",
                "required": [
                    "confirmationToken"
                ],
                "properties": {
                    "confirmationToken": {
                        "description": "Токен",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ChangePasswordRequest": {
                "title": "Запрос изменения пароля",
                "required": [
                    "currentPassword",
                    "newPassword"
                ],
                "properties": {
                    "currentPassword": {
                        "title": "Текущий пароль",
                        "description": "Класс валидации запроса редактирования пароля.",
                        "type": "string",
                        "example": "123456"
                    },
                    "newPassword": {
                        "title": "Новый пароль",
                        "description": "Класс валидации запроса редактирования пароля.",
                        "type": "string",
                        "minLength": 5,
                        "example": "123456"
                    }
                },
                "type": "object"
            },
...
            "SendEmailRequest": {
                "title": "Запрос на изменение email",
                "required": [
                    "currentPassword",
                    "email"
                ],
                "properties": {
                    "currentPassword": {
                        "title": "Текущий пароль",
                        "description": "Класс валидации запроса на изменение email.",
                        "type": "string",
                        "example": "123456"
                    },
                    "email": {
                        "title": "Электронный адрес",
                        "description": "Класс валидации запроса на изменение email.",
                        "type": "string",
                        "example": "user@terra.ru"
                    }
                },
                "type": "object"
            },
            "UpdatePlayerRequest": {
                "title": "Запрос обновления профиля пользователя",
                "properties": {
                    "playerName": {
                        "title": "Имя",
                        "description": "Класс валидации запроса обновления профиля игрока.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "title": "Пользователь",
                "properties": {
                    "id": {
                        "title": "Идентификатор пользователя.",
                        "description": "Класс JSON ответа с данными пользователя.",
                        "type": "string",
                        "example": "GWOIOISAmWFS"
                    },
                    "playerName": {
                        "title": "Игровое имя пользователя.",
                        "description": "Класс JSON ответа с данными пользователя.",
                        "type": "string",
                        "maxLength": 32,
                        "example": "Лоуренс"
                    },
                    "email": {
                        "title": "Email пользователя.",
                        "description": "Класс JSON ответа с данными пользователя.",
                        "type": "string",
                        "maxLength": 129,
                        "example": "example@gmail.com"
                    },
                    "firstName": {
                        "title": "Имя.",
                        "description": "Класс JSON ответа с данными пользователя.",
                        "type": "string",
                        "maxLength": 120,
                        "example": "Даниил."
                    },
                    "lastName": {
                        "title": "Фамилия.",
                        "description": "Класс JSON ответа с данными пользователя.",
                        "type": "string",
                        "maxLength": 120,
                        "example": "Заврин."
                    },
                    "patronymic": {
                        "title": "Отчество.",
                        "description": "Класс JSON ответа с данными пользователя.",
                        "type": "string",
                        "maxLength": 120,
                        "example": "Алексеевич."
                    },
                    "birthday": {
                        "title": "День рождения.",
                        "description": "Класс JSON ответа с данными пользователя.",
                        "type": "date",
                        "example": "1998-05-22"
                    },
                    "gender": {
                        "title": "Пол",
                        "description": "Класс JSON ответа с данными пользователя.",
                        "type": "string",
                        "enum": [
                            "male",
                            "female",
                            "other"
                        ],
                        "example": "male"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Авторизация по логину и паролю с получением токена авторизации.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}